LayoutTests/platform/chromium/fast/forms/suggestion-picker/time-suggestion-picker-mouse-operations-expected.txt

Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
PASS popupWindow.pagePopupController.toString() is "[object PagePopupController]"
Check that page popup exists.
PASS popupWindow.pagePopupController.toString() is "[object PagePopupController]"
Check that hovering over an entry highlights it.
PASS highlightedEntry() is "01:02"
Check that moving the mouse outside the popup de-highlights entries.
PASS highlightedEntry() is null
Check that mouse click closes the popup and sets the value.
PASS document.getElementById("mock-page-popup") is null
PASS document.getElementById("time").value is "01:03"
PASS successfullyParsed is true

TEST COMPLETE
